Add files via upload
As part of my MBA coursework, I conducted an in-depth comparative analysis of on-premise and cloud-based solutions for IT financial institutions. This project explored key factors such as scalability, cost-efficiency, security, and regulatory compliance to determine the optimal platform for data management and analytics. The study proposed a hybrid cloud approach that integrates the benefits of both models, ensuring enhanced security for critical data while leveraging the scalability and cost advantages of cloud platforms.

Key deliverables:

Comprehensive analysis of on-premise and cloud-based models.
Development of a decision matrix based on flexibility, cost, and compliance.
Proposal for a hybrid cloud model tailored for financial institutions.
This project provided valuable insights into modern IT infrastructure decision-making and highlighted the growing importance of hybrid cloud solutions in the financial sector.